requirements/1.txt
google-api-core[grpc]==1.22.0  # via firebase-admin, google-api-python-client, google-cloud-core, google-cloud-firestore
google-api-python-client==1.10.0  # via firebase-admin
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth==1.20.1       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.4.1  # via google-cloud-firestore, google-cloud-storage
google-cloud-firestore==1.8.1  # via firebase-admin
google-cloud-storage==1.30.0  # via firebase-admin
google-crc32c==0.1.0      # via google-resumable-media
google-resumable-media==0.7.1  # via google-cloud-storage
googleapis-common-protos==1.52.0  # via google-api-core
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests
inflection==0.5.0         # via -r requirements/base.in, drf-yasg
iniconfig==1.0.1          # via pytest
intuit-oauth==1.2.3       # via -r requirements/base.in, python-quickbooks
ipy==1.0                  # via -r requirements/base.in
isort==5.3.2              # via -r requirements/base.in
itypes==1.2.0             # via coreapi
jdcal==1.4.1              # via openpyxl
jinja2==2.11.2            # via coreschema
jmespath==0.10.0          # via boto3, botocore
jsonfield==3.1.0          # via dj-stripe
kombu==4.6.11             # via celery
markuppy==1.14            # via tablib
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via -r requirements/base.in
more-itertools==8.4.0     # via pytest
msgpack==1.0.0            # via cachecontrol
nose==1.3.7               # via docusign-esign
oauthlib==3.1.0           # via requests-oauthlib
odfpy==1.4.1              # via tablib
openpyxl==3.0.4           # via tablib
packaging==20.4           # via drf-yasg, pytest, pytest-sugar
pdfkit==0.6.1             # via -r requirements/base.in
phonenumbers==8.12.7      # via django-phonenumber-field
pilkit==2.0               # via -r requirements/base.in, django-imagekit
pillow==7.2.0             # via -r requirements/base.in
pluggy==0.13.1            # via pytest
progressbar2==3.51.4      # via django-cities-light
protobuf==3.12.4          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.5    # via -r requirements/base.in
py==1.9.0                 # via pytest, pytest-forked
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, python-jose, rsa
pycodestyle==2.6.0        # via flake8
pycparser==2.18           # via cffi, python-quickbooks
pyfcm==1.4.7              # via fcm-django
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via -r requirements/base.in
pyjwt==1.7.1              # via docusign-esign
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/base.in
pytest-django==3.9.0      # via -r requirements/base.in
pytest-env==0.6.2         # via -r requirements/base.in
pytest-forked==1.3.0      # via pytest-xdist
pytest-mock==3.2.0        # via -r requirements/base.in
pytest-sugar==0.9.4       # via -r requirements/base.in
pytest-xdist==1.34.0      # via -r requirements/base.in
pytest==6.0.1             # via -r requirements/base.in, pytest-cov, pytest-django, pytest-env, pytest-forked, pytest-mock, pytest-sugar, pytest-xdist
python-crontab==2.5.1     # via django-celery-beat
python-dateutil==2.8.1    # via arrow, awscli-cwlogs, botocore, docusign-esign, faker, freezegun, python-crontab, python-quickbooks
python-jose==3.2.0        # via intuit-oauth
python-quickbooks==0.8.2  # via -r requirements/base.in
python-utils==2.4.0       # via progressbar2
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via -r requirements/base.in, babel, celery, django, django-cities-light, django-dirtyfields, django-timezone-field, google-api-core, google-cloud-firestore
pyyaml==5.3.1             # via awscli, tablib
rauth==0.7.3              # via python-quickbooks
redis==3.5.3              # via -r requirements/base.in, django-cacheops, django-redis
requests-oauthlib==1.3.0  # via django-allauth, intuit-oauth
requests==2.24.0          # via awscli-cwlogs, cachecontrol, coreapi, django-allauth, google-api-core, intuit-oauth, pyfcm, python-quickbooks, rauth, requests-oauthlib, stripe
rest_condition==1.0.3     # via -r requirements/base.in
rsa==4.5                  # via awscli, google-auth, python-jose
sentry-sdk==0.16.3        # via -r requirements/base.in
simplejson==3.17.2        # via python-quickbooks
six==1.15.0               # via awscli-cwlogs, cryptography, django-cacheops, django-cities-light, django-coverage-plugin, django-imagekit, django-rest-auth, django-taggit-serializer, docusign-esign, drf-yasg, ecdsa, freezegun, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-resumable-media, grpcio, intuit-oauth, packaging, progressbar2, protobuf, pytest-xdist, python-dateutil, python-jose, python-quickbooks, python-utils
sqlparse==0.3.1           # via django
stripe==2.50.0            # via dj-stripe
supervisor==4.2.0         # via -r requirements/base.in
tablib[html,ods,xls,xlsx,yaml]==2.0.0  # via -r requirements/base.in, django-import-export
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest
twilio==6.60.0            # via -r requirements/base.in
unidecode==1.1.1          # via django-cities-light
uritemplate==3.0.1        # via coreapi, drf-yasg, google-api-python-client
urllib3==1.25.10          # via botocore, docusign-esign, requests, sentry-sdk
uwsgi                     # via -r requirements/base.in
vine==1.3.0               # via amqp, celery
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
django-utils-six==2.0
drf-api-tracking==1.8.0
django-easy-audit==1.3.2
django-admin-logs==1.0.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
